+++ R/rendering.R
@@ -24,6 +24,18 @@ render <- function(model,
   model_arg_label <- rlang::as_label(rlang::enexpr(model))
   if (target_tool == "nonmem") {
     issues <- check(model)
+    if ("MissingVariableIssue" %in% issue_types(issues)) {
+      if (options$issues.missing_variables %in% c("fix", "fix-warn")) {
+        model <- add_missing_variables(
+          model = model,
+          issues = issues,
+          warn = options$issues.missing_variables == "fix-warn"
+          )
+      }
+      if (options$issues.missing_variables %in% c("fix", "fix-warn", "warn", "ignore")) {
+        issues <- purrr::discard(issues, ~is(.x, "MissingVariableIssue"))
+      }
+    }
     if (length(issues) > 0) {
       rlang::abort(
         c("Critical issues found",